fixing id type + first pass at filling in ABCXodusStore, #1

+++ src/main/kotlin/org/libraryweasel/music/api/DataModel.kt
@@ -4,4 +4,4 @@
 
 package org.libraryweasel.music.api
 
-data class ABCDocument(val id: Int, val name: String, val document: String)
+data class ABCDocument(val id: Long, val name: String, val document: String)
